LayoutTests/imported/w3c/web-platform-tests/custom-elements/parser/parser-fallsback-to-unknown-element-expected.txt

CONSOLE MESSAGE: TypeError: The result of constructing a custom element must be a HTMLElement
CONSOLE MESSAGE: TypeError: The result of constructing a custom element must be a HTMLElement
CONSOLE MESSAGE: line 35: ReferenceError: Cannot access uninitialized variable.
CONSOLE MESSAGE: line 41: Bad

PASS HTML parser must create a fallback HTMLUnknownElement when a custom element constructor returns a Text node 
PASS HTML parser must create a fallback HTMLUnknownElement when a custom element constructor returns non-Element object 
PASS HTML parser must create a fallback HTMLUnknownElement when a custom element constructor does not call super() 
PASS HTML parser must create a fallback HTMLUnknownElement when a custom element constructor throws an exception 
